added the material and product api
